pass test_A030

+++ src/extargsparse/keycls.cs
@@ -313,7 +313,7 @@ public class KeyCls
                     val = (JValue) this.m_value;
                     switch(val.Type) {
                     case JTokenType.Integer:
-                        return (int) (System.Int64)(val.Value);
+                        return (System.Int64)(val.Value);
                     case JTokenType.Float:
                         return (double) val.Value;
                     case JTokenType.String:
